Moves more files. Adds new messages. Some settings fixes
