+++ src/frontend/screens/DrawerMenu.tsx
@@ -9,16 +9,16 @@ import MaterialIcon from '@react-native-vector-icons/material-icons';
 import {useNavigationFromRoot} from '../hooks/useNavigationWithTypes.ts';
 import Exchange from '../images/Exchange.svg';
 import CollaborateIcon from '../images/ProjectParticipant.svg';
-import {BodyText} from './Text/BodyText.tsx';
+import {BodyText} from '../sharedComponents/Text/BodyText.tsx';
 import {useProjectRoleAndDetails} from '../hooks/useProjectRoleAndDetails.ts';
 import {BLUE_GREY, COMAPEO_BLUE, NEW_DARK_GREY, WHITE} from '../lib/styles.ts';
 import {useActiveProject} from '../contexts/ActiveProjectContext.tsx';
-import {MenuLowStorageAlert} from './Storage/MenuLowStorageAlert.tsx';
+import {MenuLowStorageAlert} from '../sharedComponents/Storage/MenuLowStorageAlert.tsx';
 import {useStorageReadingQuery} from '../hooks/useStorageReadingQuery.ts';
-import {ColorCard} from './ColorCard.tsx';
-import {HeaderText} from './Text/HeaderText.tsx';
+import {ColorCard} from '../sharedComponents/ColorCard.tsx';
+import {HeaderText} from '../sharedComponents/Text/HeaderText.tsx';
 import {useManyProjects} from '@comapeo/core-react';
-import {buttonStyles, PrimaryButton} from './Buttons.tsx';
+import {buttonStyles, PrimaryButton} from '../sharedComponents/Buttons.tsx';
 import DownArrow from '../images/DownArrow.svg';
 import {isLowStorage, calcUsedPercentage} from '../lib/storage';
 import {useEarlyAccessState} from '../contexts/EarlyAccessContext';
